Sped up calculations
It's basically limited by the framerate, and I'm not sure if I could switch to threads instead of coroutines to make it faster, or if it would even be worth it to attempt switching.
There's still bugginess with newly downloaded songs, and idk if there's anything I can really do about that.