Improved lag spikes, not completely gone though
I think removing the use of GetFrameResults is going to be necessary to eliminate the lag spikes.
Currently I use it to get info about the latest note hits, and setting the score properly on saving and on the results screen.
If I can find alternatives for both of these, the lag spikes should be gone. I should be able to copy over what I use in DaniDojo for reading hit notes.